% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_race_octal.R
\name{decode_race_name}
\alias{decode_race_name}
\title{Decode a race code to a pathotype vector}
\usage{
decode_race_name(code, system = c("octal", "decanary"), n = NULL)
}
\arguments{
\item{code}{A single race code.}

\item{system}{One of \code{"octal"} or \code{"decanary"}.}

\item{n}{Optional integer length of the original pathotype vector.}
}
\value{
An integer vector of 0/1 values.
}
\description{
Decode a race code to a pathotype vector
}
