% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.as_data_table}
\alias{.as_data_table}
\title{Coerce Input to a data.table}
\usage{
.as_data_table(.x)
}
\arguments{
\item{.x}{an object to coerce}
}
\value{
a \code{data.table}, copied if \code{.x} was already one
}
\description{
Coerce Input to a data.table
}
\keyword{internal}
