% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.binary_cutoff}
\alias{.binary_cutoff}
\title{Create Binary Reaction Value}
\usage{
.binary_cutoff(.x, .cutoff)
}
\arguments{
\item{.x}{A \code{data.table} containing the values to be summarised}

\item{.cutoff}{Cutoff value for susceptibility}
}
\value{
A \code{data.table} that tallies the results by gene
}
\description{
Adds a column of 1 or 0 for the susceptible reaction cutoff
}
\keyword{internal}
