% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.check_inputs}
\alias{.check_inputs}
\title{Check User Inputs}
\usage{
.check_inputs(.x, .cutoff, .control, .sample, .gene, .perc_susc)
}
\arguments{
\item{.x}{a \code{data.table} containing the values to be summarised}

\item{.cutoff}{value for percent susceptible cutoff. Numeric.}

\item{.control}{value used to denote the susceptible control in the \code{gene}
column. Character.}

\item{.sample}{column providing the unique identification for each sample
being tested. Character.}

\item{.gene}{column providing the gene(s) being tested. Character.}

\item{.perc_susc}{column providing the percent susceptible reactions as a
numeric value. Character.}
}
\description{
Checks and validates user inputs before running functions
}
\keyword{internal}
