% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_race_octal.R
\name{.decanary_string_to_value}
\alias{.decanary_string_to_value}
\title{Parse and Validate a Decanary Code String's Numeric Value}
\usage{
.decanary_string_to_value(code)
}
\arguments{
\item{code}{a single character string representing a non-negative integer}
}
\value{
a single integer value
}
\description{
Parse and Validate a Decanary Code String's Numeric Value
}
\keyword{internal}
