% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_race_octal.R
\name{.decanary_to_bits_fixed}
\alias{.decanary_to_bits_fixed}
\title{Decode a Decanary Value to a Fixed Number of Bits}
\usage{
.decanary_to_bits_fixed(value, n)
}
\arguments{
\item{value}{a single non-negative integer}

\item{n}{the number of bits to return}
}
\value{
an integer vector of \code{n} 0/1 values
}
\description{
Decode a Decanary Value to a Fixed Number of Bits
}
\keyword{internal}
