% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_race_octal.R
\name{.decode_octal_digits}
\alias{.decode_octal_digits}
\title{Decode a Vector of Octal Digits to Binary Bits}
\usage{
.decode_octal_digits(digits)
}
\arguments{
\item{digits}{an integer vector of octal digits (0-7)}
}
\value{
an integer vector of 0/1 values, 3 bits per digit
}
\description{
Decode a Vector of Octal Digits to Binary Bits
}
\keyword{internal}
