% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_race_octal.R
\name{.decode_one_octal_digit}
\alias{.decode_one_octal_digit}
\title{Decode a Single Octal Digit to its Three Binary Bits}
\usage{
.decode_one_octal_digit(d)
}
\arguments{
\item{d}{a single octal digit (0-7)}
}
\value{
an integer vector of length 3 (0/1 values)
}
\description{
Decode a Single Octal Digit to its Three Binary Bits
}
\keyword{internal}
