% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_race_octal.R
\name{.octal_string_to_digits}
\alias{.octal_string_to_digits}
\title{Split an Octal Code String into Validated Integer Digits}
\usage{
.octal_string_to_digits(code)
}
\arguments{
\item{code}{a single character string of octal digits 0-7}
}
\value{
an integer vector of octal digits
}
\description{
Split an Octal Code String into Validated Integer Digits
}
\keyword{internal}
