% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.rename_input_columns}
\alias{.rename_input_columns}
\title{Validate and Rename the Requested Columns}
\usage{
.rename_input_columns(dt_x, .perc_susc, .gene, .sample)
}
\arguments{
\item{dt_x}{a \code{data.table}}

\item{.perc_susc, .gene, .sample}{column names to look for in \code{dt_x}}
}
\value{
\code{dt_x} with the requested columns renamed
}
\description{
Checks that \code{.perc_susc}, \code{.gene}, and \code{.sample} exist in \code{dt_x}, then
renames them to \code{perc_susc}, \code{gene}, and \code{sample} respectively.
}
\keyword{internal}
