% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.validate_input_types}
\alias{.validate_input_types}
\title{Validate the Types of \code{.check_inputs()} Arguments}
\usage{
.validate_input_types(.x, .control, .sample, .gene, .perc_susc)
}
\arguments{
\item{.x}{a \code{data.table} containing the values to be summarised}

\item{.control}{value used to denote the susceptible control in the \code{gene}
column. Character.}

\item{.sample}{column providing the unique identification for each sample
being tested. Character.}

\item{.gene}{column providing the gene(s) being tested. Character.}

\item{.perc_susc}{column providing the percent susceptible reactions as a
numeric value. Character.}
}
\value{
Invisibly \code{TRUE}, or throws an error.
}
\description{
Validate the Types of \code{.check_inputs()} Arguments
}
\keyword{internal}
