% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.validate_perc_susc}
\alias{.validate_perc_susc}
\title{Validate the perc_susc Column}
\usage{
.validate_perc_susc(dt_x)
}
\arguments{
\item{dt_x}{a \code{data.table} already renamed by \code{.rename_input_columns()}}
}
\value{
Invisibly \code{TRUE}, or throws an error.
}
\description{
Validate the perc_susc Column
}
\keyword{internal}
