% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_race_octal.R
\name{encode_race_decanary}
\alias{encode_race_decanary}
\title{Encode a binary pathotype vector as reverse binary/decanary}
\usage{
encode_race_decanary(x)
}
\arguments{
\item{x}{Integer/numeric/logical vector of 0/1 values.}
}
\value{
A single character string containing the reverse decanary race name.
}
\description{
Encode a binary pathotype vector as reverse binary/decanary
}
\keyword{internal}
