% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_diversities.R
\name{print.hagis.diversities}
\alias{print.hagis.diversities}
\title{Prints hagis.diversities Object}
\usage{
\method{print}{hagis.diversities}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{a \code{hagis.diversities} object}

\item{digits}{number of digits to print}

\item{...}{ignored}
}
\description{
Custom \code{\link[=print]{print()}} method for \code{hagis.diversities} objects.
}
