% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virulence_difference_matrix.R
\name{virulence_difference_table}
\alias{virulence_difference_table}
\title{Calculate Pairwise Virulence-Difference Table}
\usage{
virulence_difference_table(x)
}
\arguments{
\item{x}{A binary matrix, usually produced by \code{\link[=create_binary_matrix]{create_binary_matrix()}}.}
}
\value{
A \code{data.table} with one row per pairwise comparison.
}
\description{
Calculate Pairwise Virulence-Difference Table
}
