% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_complexities.R
\name{.create_summary_isolate}
\alias{.create_summary_isolate}
\title{Create Summary Table of Binary Reactions by Sample}
\usage{
.create_summary_isolate(.y)
}
\arguments{
\item{x}{A \CRANpkg{hagis} \code{complexities} object generated by
\code{\link[=calculate_complexities]{calculate_complexities()}}. \code{Character}.}
}
\value{
A \code{data.table} that tallies the results by sample.
}
\description{
Tally a summary by sample or isolate. This code takes the "Susceptible.1"
column and summarises it by gene for your total "Isolates" pathogenic on
each sample.
}
\keyword{internal}
